% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ggimage}
\alias{ggimage}
\alias{get_out}
\title{plot and get result of run_model}
\usage{
get_out(out, which_var, depth = NULL)

ggimage(
  out,
  which_var = "Biomass",
  trans = NULL,
  y.rev = TRUE,
  add.contour = FALSE,
  contour.text = FALSE,
  depth = NULL,
  sign_var = FALSE
)
}
\arguments{
\item{out}{Return of \link{ode.1D}}

\item{which_var}{which variable in \code{attributes(out)}}

\item{depth}{depth name for variable}

\item{trans}{Transformation of \link{scale_fill_gradientn}, like \code{"log10"}.}

\item{y.rev}{Should be y scale on the reverse scale?}

\item{contour.text}{Should contour text be printed?}

\item{sign_var}{Does variable have sign attribute?}
}
\value{
A matrix

ggplot list
}
\description{
plot and get result of run_model
}
