% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run AlgalGame model}
\usage{
run_model(
  inputs = NULL,
  Ttot = 2,
  dt = 0.001,
  L = 2,
  N = 50,
  ini = NULL,
  opt.abg = "Zhou2019",
  plot = TRUE,
  run.steady = FALSE,
  run.steady.list = list(time = 0),
  moving_avg = FALSE
)
}
\arguments{
\item{inputs}{A list of parameters. Run \link{show_parms} to see all settings.}

\item{Ttot}{Total of time}

\item{dt}{Time interval}

\item{L}{Column depth}

\item{N}{Number of vertical grids}

\item{ini}{Initial states of Biomass and Nutrient}

\item{opt.abg}{option of abg, default as \code{"Zhou2019"}}

\item{plot}{whether to plot the raster}

\item{out_steady}{Whether to run steady resolver}

\item{out_steady.list}{Input params of \link{steady.1D}}
}
\value{
An out object.
}
\description{
Run AlgalGame model
}
\examples{
library(AlgalGame)
res <- run_model()
ggimage(res$out)

}
\references{
Klausmeier C A, Litchman E. Algal games: The vertical distribution of phytoplankton in
poorly mixed water columns[J]. Limnology and Oceanography, 2001, 46(8): 1998-2007.
}
